moves TemporaryPanelTabPane into its own file
